Updated create ripple app for latest changes
