+++ packages/create-ripple-app/template/nuxt/nuxt.config.js
@@ -50,30 +50,30 @@ export default {
   ** Build
   * https://nuxtjs.org/api/configuration-build/
   */
- build: {
-  extend (config, { isDev, isClient }) {
-    const webpack = require('webpack')
-    const LodashModuleReplacementPlugin = require('lodash-webpack-plugin')
-    config.plugins.push(new LodashModuleReplacementPlugin({
-      'caching': true,
-      'collections': true,
-      'paths': true,
-      'shorthands': true
-    }))
-    // Load moment 'en-au' locale only for performance.
-    // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-    // You need to change it if your site is not in Australia.
-    config.plugins.push(new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /en-au/))
-  },
+  build: {
+    extend (config, { isDev, isClient }) {
+      const webpack = require('webpack')
+      const LodashModuleReplacementPlugin = require('lodash-webpack-plugin')
+      config.plugins.push(new LodashModuleReplacementPlugin({
+        'caching': true,
+        'collections': true,
+        'paths': true,
+        'shorthands': true
+      }))
+      // Load moment 'en-au' locale only for performance.
+      // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
+      // You need to change it if your site is not in Australia.
+      config.plugins.push(new webpack.ContextReplacementPlugin(/moment[/\\]locale$/, /en-au/))
+    },
 
-  // Currently lodash is mainly brought by Elastic search JS lib.
-  // Below lodash optimization can be reviewed after we migrate to new ES JS client.
-  babel: {
-    plugins: [
-      'lodash'
-    ]
-  }
-},
+    // Currently lodash is mainly brought by Elastic search JS lib.
+    // Below lodash optimization can be reviewed after we migrate to new ES JS client.
+    babel: {
+      plugins: [
+        'lodash'
+      ]
+    }
+  },
   css: [
     '@/assets/_custom.scss'
   ],
